refactor(cli): colocate session commands
